aligned implementation with specs
