+++ pyproject.toml
@@ -25,6 +25,7 @@ dev = [
 tests = [
     "pytest>=7.1.2",
     "pytest-cov>=4.0.0",
+    "pytest-subprocess>=1.5.0",
 ]
 oarepo14 = [
     # just for installation, no code dependency
