+++ tests/unit/__init__.py
@@ -1,0 +1,4 @@
+# SPDX-FileCopyrightText: 2026 CESNET z.s.p.o.
+# SPDX-License-Identifier: MIT
+
+from __future__ import annotations
